Removed CGAL library as longer needer
Previously used Voronoi region based roadmap generation was replaced with the need floodfill strategy working on a grid.  The new apporach is more robust as it can incorporate any shade and doesnt have the limitation of working only with points as the previous apporach. It is speculated that the floodfill strategy will be also faster